consolidate test plans for upgrades
