+++ cmd/action/update/cluster/version/error.go
@@ -1,4 +1,4 @@
-package patch
+package version
 
 import "github.com/giantswarm/microerror"
