+++ cmd/action/update/cluster/version/runner.go
@@ -1,4 +1,4 @@
-package patch
+package version
 
 import (
 	"context"
