+++ cmd/plan/command.go
@@ -10,8 +10,6 @@ import (
 	"github.com/giantswarm/awscnfm/v12/cmd/plan/pl003"
 	"github.com/giantswarm/awscnfm/v12/cmd/plan/pl004"
 	"github.com/giantswarm/awscnfm/v12/cmd/plan/pl005"
-	"github.com/giantswarm/awscnfm/v12/cmd/plan/pl006"
-	"github.com/giantswarm/awscnfm/v12/cmd/plan/pl007"
 )
 
 const (
@@ -90,30 +88,6 @@ func New(config Config) (*cobra.Command, error) {
 		}
 	}
 
-	var pl006Cmd *cobra.Command
-	{
-		c := pl006.Config{
-			Logger: config.Logger,
-		}
-
-		pl006Cmd, err = pl006.New(c)
-		if err != nil {
-			return nil, microerror.Mask(err)
-		}
-	}
-
-	var pl007Cmd *cobra.Command
-	{
-		c := pl007.Config{
-			Logger: config.Logger,
-		}
-
-		pl007Cmd, err = pl007.New(c)
-		if err != nil {
-			return nil, microerror.Mask(err)
-		}
-	}
-
 	f := &flag{}
 
 	r := &runner{
@@ -135,8 +109,6 @@ func New(config Config) (*cobra.Command, error) {
 	c.AddCommand(pl003Cmd)
 	c.AddCommand(pl004Cmd)
 	c.AddCommand(pl005Cmd)
-	c.AddCommand(pl006Cmd)
-	c.AddCommand(pl007Cmd)
 
 	return c, nil
 }
